backend/requirements.txt
Django==5.0.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
mysqlclient==2.2.4
openpyxl==3.1.5
python-dotenv==1.0.1
argon2-cffi==23.1.0
django-ratelimit==4.1.0
whitenoise==6.7.0
gunicorn==22.0.0
